Refactored the two country property editors into one.
They were almost identical, so it made sense.

At the same time, the component was split into three parts - viewer,
editor, and top-level controller. When you have a component that
is completely different based on some state, it doesn't make much
sense for it to be one component, it's just harder to read that way.

This also fixed a bug where hitting enter when editing the URL reloaded
the page.