Removed some redundancy in General Info
When you see something repeating more than 2-3 times, it's probably
worth it to remove the redundancy, especially when it's as easy as here.
Compare readability between the two..

Also, creating a function like this renderTable(), which simply renders
a table, and is called once, almost immediately after being declared,
kinda doesn't serve any purpose, it just adds noise.

Also, createRow didn't have to be inside the component (getting
recreated all the time), so it was moved out.